Better task importing
