+++ flight/drone/drone_controller.py
@@ -9,8 +9,8 @@ import traceback
 from drone import Drone
 import exceptions
 
-import flight.constants as c
-import flight.tasks as tasks
+from flight import constants as c
+import flight.tasks
 from flight.utils.priority_queue import PriorityQueue
 from flight.utils.timer import Timer
 from flight import flightconfig as f
@@ -130,7 +130,7 @@ class DroneController(object):
         priority : Priorities.{LOW, MEDIUM, HIGH}, optional
             The importance of this task.
         """
-        new_task = tasks.Hover(self._drone, altitude, duration)
+        new_task = flight.tasks.Hover(self._drone, altitude, duration)
         self._task_queue.push(priority, new_task)
 
     def add_takeoff_task(self, altitude, priority=c.Priorities.HIGH):
@@ -147,7 +147,7 @@ class DroneController(object):
         -----
         Internally, the priority of this task is always set to HIGH.
         """
-        new_task = tasks.Takeoff(self._drone, altitude)
+        new_task = flight.tasks.Takeoff(self._drone, altitude)
         self._task_queue.push(priority, new_task)
 
     def add_linear_movement_task(
@@ -163,7 +163,7 @@ class DroneController(object):
         priority : Priorities.{LOW, MEDIUM, HIGH}, optional
             The importance of this task.
         """
-        new_task = tasks.LinearMovement(self._drone, direction, duration)
+        new_task = flight.tasks.LinearMovement(self._drone, direction, duration)
         self._task_queue.push(priority, new_task)
 
 
@@ -175,7 +175,7 @@ class DroneController(object):
         priority : Priorities.{LOW, MEDIUM, HIGH}, optional
             The importance of this task.
         """
-        new_task = tasks.Land(self._drone)
+        new_task = flight.tasks.Land(self._drone)
         self._task_queue.push(priority, new_task)
 
     def add_exit_task(self, priority=c.Priorities.HIGH):
@@ -185,7 +185,7 @@ class DroneController(object):
         -----
         Always has high priority
         """
-        new_task = tasks.Exit(self._drone)
+        new_task = flight.tasks.Exit(self._drone)
         self._task_queue.push(priority, new_task)
 
     def _update(self):
@@ -205,7 +205,7 @@ class DroneController(object):
                 # We are done with the task
                 self._logger.info('Finished {}...'.format(
                     type(self._current_task).__name__))
-                if isinstance(self._current_task, tasks.Exit):
+                if isinstance(self._current_task, flight.tasks.Exit):
                     return False
                 self._task_queue.pop()
